pyproject.toml
[tool.poetry]
name = "DB PROJECT"
version = "0.1.0"
description = "database project"
authors = ["group l"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.2"
mysql-connector-python = "^8.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
